backend/requirements-jobs.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements-jobs.txt requirements-jobs.in
#
cachecontrol==0.12.11
    # via -r common/requirements.in
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via flask
flask==2.1.3
    # via -r common/requirements.in
google-api-core[grpc]==2.8.2
    # via
    #   google-api-python-client
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-api-python-client==2.54.0
    # via -r requirements-jobs.in
google-auth==2.9.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-aiplatform==1.16.1
    # via -r requirements-jobs.in
google-cloud-appengine-logging==1.1.3
    # via google-cloud-logging
google-cloud-audit-log==0.2.3
    # via google-cloud-logging
google-cloud-bigquery==2.34.4
    # via
    #   -r requirements-jobs.in
    #   google-cloud-aiplatform
google-cloud-core==2.3.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==3.2.1
    # via -r common/requirements.in
google-cloud-pubsub==2.13.4
    # via -r common/requirements.in
google-cloud-resource-manager==1.6.0
    # via google-cloud-aiplatform
google-cloud-storage==2.4.0
    # via
    #   -r requirements-jobs.in
    #   google-cloud-aiplatform
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.56.4
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.4
    # via
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   google-cloud-resource-manager
grpcio==1.47.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.47.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
gunicorn==20.1.0
    # via -r common/requirements.in
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via jinja2
msgpack==1.0.4
    # via cachecontrol
packaging==21.3
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
proto-plus==1.20.6
    # via
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   google-cloud-resource-manager
protobuf==3.20.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-bigquery
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.9
    # via
    #   httplib2
    #   packaging
python-dateutil==2.8.2
    # via google-cloud-bigquery
requests==2.28.1
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   grpcio
    #   python-dateutil
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.10
    # via requests
werkzeug==2.0.3
    # via flask
zipp==3.8.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
